Issue #4 - implement floating to fixed point and find best scaling value.

+++ gemm_fp/float_fixed.cpp
@@ -3,14 +3,14 @@
 
 int roundup(float fp_number)
 {
-	int	fx_number	=	(int)fp_number;
+	long long int	fx_number	=	(long long int)fp_number;
 
 	if(fp_number-fx_number>=0.5)	fx_number++;
 
 	return	fx_number;
 }
 
-void mm_float_to_fixed (float* A_float, int* A_fixed, int M, int K, int scale) {
+void mm_float_to_fixed (float* A_float, long long int* A_fixed, int M, int K, int scale) {
 	int length = M*K;
 
 	for (int i=0; i<length; i++) {
@@ -18,9 +18,8 @@ void mm_float_to_fixed (float* A_float, int* A_fixed, int M, int K, int scale) {
 	}
 }
 
-
 /** Fixed to floating point */
-void mm_fixed_to_float (int* A_fixed, float* A_float, int M, int K, int scale) {
+void mm_fixed_to_float (long long int* A_fixed, float* A_float, int M, int K, int scale) {
 	int length = M*K;
 
 	for (int i=0; i<length; i++) {
